requirements.txt
fastapi
uvicorn
httpx
beautifulsoup4
feedparser
python-dotenv
pydantic